fix: Fix crash on older browser versions
Fix crash on older browser versions caused by broken polyfill. Affected
browser versions were Chrome versions 109-118, and Firefox versions
102-120.

The error was caused by the `withResolvers` polyfill. The polyfill
tried to add the `withResolvers` function to the `Promise` global if it
was missing, but this is forbidden under SES lockdown and would fail in
a production build.

I worked around this problem by converting the polyfill to a ponyfill
that had to be imported, rather than mutating a global. This should
work the same way but avoids the lockdown error.

Fixes #34235