fix: Fix silently dropped errors in UI
Update all UI Sentry calls to use `global.Sentry` instead of
imported library functions.

The imported functions do not work in production builds because they
refer to a copy of that code run in a Compartment (by LavaMoat), which
is never initialized. The Sentry client we initialize is not in a
compartment.