chore: Add buttons to test error capture
Add two buttons to the "Developer Options" settings page to test that
captured Sentry errors work correctly (one button for the UI, one for
the background).

We have two similar buttons already for unhandled errors, but not for
explicitly handled errors.

These will assist with manual testing for this bug: #34371