fix: Fix frequent writes from CronjobController
Prevent frequent high-volume writes caused by state updates from the
CronjobController. The controller state includes an event with a
timestamp that can be updated very frequently if the user has a snap
installed with a scheduled recurring event with a short interval. It
can even run while the wallet is otherwise idle, causing the entirety
of wallet state to be written to disk repeatedly.

We have temporarily mitigated the problem by moving the
CronjobController state to a new top-level storage key, so that it
can be written to independently of other state. This controller's state
is very small and should not cause any significant amount of write
volume.