fix: Stop polling multichain rates while idle
The `MultichainRatesController` now stops polling when the UI is
closed, and starts up again when the first UI instance is opened
(conditionally, if a non-EVM account is selected).

The method for starting polling is slightly different from other
polling controllers because this controller doesn't use the same
polling mixin as elsewhere. This controller is deprecated and soon to
be removed.

Relates to #33879