test: Fix flaky account sync test
This account sync test was flaky because it wasn't properly waiting for
all account data to finish synchronizing. It was waiting for a single
sync event rather than two. It's unclear why two events are expected,
when this changed (if it did change), or why this wasn't flaky before.
But this seems to make the test consistently pass.

Adding a delay after this point also results in a consistently passing
test, so it's like not a functional bug (since it works if we wait for
snyncing to complete).