chore: Remove Sentry library from main bundles
The Sentry library is setup in a dedicated `sentry-install.js` script
that is run before our main bundles. However, it was still present in
our main bundles because of a log utility that was imported from a
module used by `sentry-install.js`.

The log utility has been moved to a shared function to ensure that the
bulk of the `sentry-install.js` code is not redundantly included in the
main bundles as well. This also helps prevent regressions on #34371 by
ensuring any accidental re-introduction of the Sentry library in the
main bundle is highlighted by a LavaMoat policy change.