Improve tracking of RPC endpoint failures
Currently, we only track when an Infura RPC endpoint becomes degraded or
unavailable. Now, we would like to have similar insights about custom
RPC endpoints so that we can take more informed decisions to improve
reliability for other chains. We'd also like to improve the tracking for
Infura endpoints so that we can understand failures better.

This commit updates the handlers for the
`NetworkController:rpcEndpointDegraded` and
`NetworkController:rpcEndpointUnavailable` messenger events so that they
create a Segment event regardless of the type of endpoint. The event now
includes the HTTP status code.

While making these changes, it was noticed that the sampling rate for
these Segment event was incorrect. It should have been 1%, not 10%. This
has also been corrected. This ensures that we don't store more data in
Segment and our downstream services than necessary.