test: Fix console error in onboarding E2E tests
The E2E tests using the onboarding fixture would consistently show a
background console error about metadata being missing for the
`skipDeepLinkInterstitial` property because it was accidentally added
to the root of the `PreferencesController` state object rather than in
the `preferences` object. It shouldn't really have been in the state
fixture at all because the fixture's schema is meant to match migration
version 74, which is quite old.

The error was resolved by removing the property.