test: Optimize test server shutdown
The `withFixtures` helper function has been updated to shut down test
servers in parallel rather than one-at-a-time. Any shutdown errors are
logged, then the first error is re-thrown.

Potentially this could obscure the underlying test error (if there was
one), but this seems like an acceptable tradeoff for ensuring bugs in
the shutdown code are discovered and prioritized.