test: Fix flaky test failure upon extension reset
When an E2E test causes the extension to reset, it can sometimes cause
the webdriver to shut down automatically when it loses its connection
to the browser. When this happens, the `driver.quit` call throws during
the shutdown process. This causes the test to fail, and it leaves later
test servers running which causes later tests to fail as well.

The `driver.quit` method has been updated to no longer throw an error
when the driver is already shut down.

Fixes #34739