test: Support window switching without mocha server
Our E2E driver class now allows the methods `switchWindow` and
`switchToWindowWithTitle` methods to be used in tests that have the
Mocha background server disabled (such as those that use a production
build, or those that involve the extension resetting).

Previously the `switchToWindowWithTitleWithoutSocket` driver method
served this purpose, but the test author had to know to use this
alternative method. Plus it didn't allow switching to a window handle,
or any of the other switch methods that internally use `switchWindow`.

There are a couple of switch methods still not supported, but in these
cases an error will be thrown so that the failure reason will be clear.

This unblocks #31435