fix: Add missing build secret to Firefox bundle script
The Firefox bundle script was missing the `SENTRY_DSN_DEV` secret. We
had to add it manually the last time we submitted to Firefox.

It was omitted originally because we didn't realize it was embedded in
the build. It gets included in production builds despite not actually
being used.

Fixes #35279