feat: add error message for unsupported Ethereum NFTs on Ledger Nano S
This commit introduces a new error message for users attempting to send or manage Ethereum NFTs on the Ledger Nano S, informing them of the device's limitations and providing a link to upgrade options. Additionally, the transaction failure handling in the extension has been updated to display this new message when applicable.