chore: Disable React & Redux Devtools by default
When running `yarn start` or `yarn webpack --watch` and not running
`yarn devtools:react` or `yarn devtools:redux`, noisy WebSocket errors
appear in the browser DevTools console window from React and Redux
DevTools.

Investigation revealed that we seem to disable Redux DevTools unless
`METAMASK_DEBUG` is set to `true`, but we do not do the same thing for
React DevTools.

However, loading React DevTools only when `METAMASK_DEBUG` is set to
`true` poses another problem, which is that `METAMASK_DEBUG` is always
set to `true` in development and can't be disabled.

To address the above problem, this commit introduces a new environment
variable, `METAMASK_REACT_REDUX_DEVTOOLS`, and uses it to wrap loading
of React and Redux DevTools instead of `METAMASK_DEBUG`. It also
upgrades React DevTools to the latest version so that we can wrap it
effectively.