chore: Don't wait for net_version before starting UI
When starting, the UI creates a StreamProvider that wraps the MetaMask
provider and can be used to make internal and external requests through
the core JSON-RPC pipeline via the background connection. This
StreamProvider must be initialized in order for the UI to complete the
start sequence and in order for users to see anything on screen.

When the StreamProvider initializes, it will make a background request
for `metamask_getProviderState` to populate its internal state. As a
part of its logic, this RPC method will retrieve the network version
from whichever RPC endpoint is currently selected. However, this creates
a problem, because if the network is slow, this request won't resolve in
a timely manner, causing the UI will be stuck in a loading state and for
users to see a spinner in the meantime.

To address this, this commit changes the `metamask_getProviderState`
handler to accept a parameter, `isInitializingStreamProvider`. It
expects that StreamProvider will call the RPC method and pass `true` for
this parameter. When this happens, the request for `net_version` will
not be made and the initial network version will be set to "loading".
This allows the UI to complete its start sequence. Once the UI is
loaded, it will update the `memStore` state, and
`metamask_getProviderState` will be requested again, but this time,
`isInitializingStreamProvider` will be set to `false`. This means the
network version will be properly retrieved and propagated to all of the
places that need it.