test: Update test splitting to support new test workflows
The script responsible for determining how to split up tests in a given
workflow was failing when I tried to add a new workflow. It didn't seem
to handle the case where a workflow had zero past runs.

It has been updated to use a naive fallback splitting algorithm in this
case.