chore: Update `eth-snap-keyring` and `keyring-api`
The packages `@metamask/eth-snap-keyring` and `@metamask/keyring-api`
have been updated to their latest versions. This was done because the
current version of `@metamask/eth-snap-keyring` we're using has outdated
version of `@metamask/base-controller`, but this update includes a bump
to the latest major version.

Changelogs:
- `@metamask/eth-snap-keyring`: https://github.com/MetaMask/accounts/blob/main/packages/keyring-snap-bridge/CHANGELOG.md#1700
  - Just the `@metamask/keyring-api` peer dependency bump
- `@metamask/keyring-api`: https://github.com/MetaMask/accounts/blob/main/packages/keyring-api/CHANGELOG.md#2100
  - The `BtcMethod.SendBitcoin` method was removed, but this was
handled already in #35404 (the breaking changes were mistakenly
introduced as a patch update first, which is already updated here)

Relates to https://github.com/MetaMask/core/issues/5626