Replace "network connecting spinner" with skeleton view
Currently, the user will see a spinner when the globally selected
network is connecting. If the connection is confirmed, then the spinner
disappears, but if not, then a "connection failed" modal will appear and
ask the user to try again or switch networks.

However, this is not a great user experience. The main issue is that the
spinner blocks the majority of the UI, preventing users from switching
networks while the connection is being confirmed. Besides this, it is
incompatible with the other changes we've been gradually rolling out to
enable users to interact with multiple chains at once, as it relies on
the outdated concept of the globally selected network.

To work toward addressing these issues, this PR removes the "connection
loading" spinner and "connection failed" modal altogether. Instead,
when the user is loading the home screen for the first time — and thus,
we have not confirmed connection for any of the networks that are
currently enabled — they will see skeleton loaders in place of the total
balance and in place of balances for tokens that we've loaded so far. As
soon as one enabled network is confirmed, then the skeletons will
disappear and the user will start seeing balances come in. Balances and
the token list may update as more data is retrieved.

In addition, each time the user loads or returns to the home screen or
enables a new set of networks, the connection status for the newly
enabled networks will be checked again. This logic is not perfect —
sometimes the network status is cached and therefore may be inaccurate —
but it is better than what we do currently, which is to only check the
network status once, when the extension starts.

Note that this PR does not address what happens if an enabled network is
degraded or unavailable. That will be handled in another PR.