chore: Update `assets-controllers` patch comment
I spent some time recently investigating why we need this patch. The
current comment is a bit misleading, as the issue has nothing to do
with Jest and the problem is not just ESM support.

The comment has been expanded to explain the issue a bit better, link
further context, and describe when we can attempt deleting the patch
again.