test: fix flaky test `Test Snap getEntropy can use snap_getEntropy inside a snap` (#36087)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**
The test is flaky due to a limitiation of the window switch whenever
there is the alert open (modal dialogs, such as alerts, confirms,
prompts, are blocking by design). The normal (and working flow) is:

1. We trigger a function from the test dapp
2. MetaMask dialog opens
3. We confirm, and MetaMAsk dialog closes
4. We switch back to the test dapp
5. After some seconds, a browser alert appears

However, if the alert(5) appears *before* we have switched windows, then
it's not possible to switch windows and we get into a limbo state -->
the dialog is closed, and we cannot switch to the test dapp to
accept/close the alert.

In this case, an error appears and the test fails. I found that no
matter if we ignore the error, the test will fail after because we try
to close the dialog (not possible bc we cannot switch windows), so the
best solution is to modify the waitForBrowserAlert function to be more
robust. This way, we'll be able to use it in all cases (both in the case
we are already in the correct window where the alert will appear, and in
the case where we are *not* in the window where the alert will appear).

<img width="1190" height="182" alt="image"
src="https://github.com/user-attachments/assets/69c3a1e7-e076-4783-81b3-6fb730b1805c"
/>

<!--
Write a short description of the changes included in this pull request,
also include relevant motivation and context. Have in mind the following
questions:
1. What is the reason for the change?
2. What is the improvement/solution?
-->

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/36087?quickstart=1)

## **Changelog**

<!--
If this PR is not End-User-Facing and should not show up in the
CHANGELOG, you can choose to either:
1. Write `CHANGELOG entry: null`
2. Label with `no-changelog`

If this PR is End-User-Facing, please write a short User-Facing
description in the past tense like:
`CHANGELOG entry: Added a new tab for users to see their NFTs`
`CHANGELOG entry: Fixed a bug that was causing some NFTs to flicker`

(This helps the Release Engineer do their job more quickly and
accurately)
-->

CHANGELOG entry:

## **Related issues**

Fixes:

## **Manual testing steps**

1. Go to this page...
2.
3.

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**

<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [ ] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask
Extension Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.