fix: Sidestep rounding errors with Jazzicons
We use the `@metamask/jazzicon` library to render account avatars. This
library generates an SVG icon by generating a random number and then
performing math to determine how to lay out the parts of the icon.
However, the coordinates of these parts come out to numbers with a large
number of decimal numbers, and due to floating point rounding errors,
the number of decimals may vary each time the library is used. This
impacts snapshot tests for components that make use of `jazzicon`
because it means that sometimes the generated SVG fails to match the
snapshot.

To address the issue and remove the intermittent failures, this commit
patches `jazzicon` to limit the number of decimal numbers in the
generated SVG code. We will fix `jazzicon` directly in a future update.