Notify users about degraded and unavailable networks
When the user loads the home screen for the first time or changes the
set of enabled networks, we will attempt to verify that requests to
those networks are successful, thereby judging whether those networks
are available. If verification for any of the enabled networks takes 5
seconds or longer, then we will notify the user by popping up a banner,
mentioning which network is degraded and advising that the user update
the default RPC endpoint for that network. If verification takes 30
seconds or longer, then we will assume that the network is unavailable
and inform the user that the connection failed.

We also track metrics for when a user sees these banners and when the
user chooses to update the RPC.