chore: Decrease threshold before RPC failover is activated
Currently, an RPC endpoint that is configured with a failover endpoint
must have 28 consecutive failed attempts before the failover will be
activated. Since this also controls when requests to the primary are
temporarily paused, this threshold was intentionally kept high to
prevent accidental activations, but now we are seeing that if the
endpoint is merely degraded but not completely down — sometimes
returning non-200s but not always — then the failover will never be
activated. This commit lowers the threshold so we can handled degraded
situations more gracefully.