fix: Decrease RPC failover threshold & cooldown duration
Currently, there must be 28 consecutive failed attempts to hit an RPC
endpoint before it is perceived to be unavailable. At this point, if the
endpoint is configured with a failover, the failover will be activated
and all requests will be automatically diverted to it; otherwise
requests are paused for 30 minutes.

There are two problems we are seeing now:

- If Infura is degraded enough to where failing over to QuickNode is
  warranted, the failover may be activated too late.
- If the user is attempting to access a custom network and they are
  experiencing issues (either due to a local connection issue or an
  issue with the endpoint itself), they will be prevented from using
  that network for 30 minutes. This is way too long.

To fix these problems, this commit:

- Lowers the "max consecutive failures" (the number of successive
  attempts to obtain a successful response from an endpoint before
  requests are paused or the failover is triggered) from 28 to 8.
- Lowers the "circuit break duration" (the period during which requests
  to an unavailable endpoint will be paused) from 30 minutes to 30
  *seconds*.

In summary, if a network starts to become degraded or the user is
experiencing connection issues, the network is more likely to be flagged
as unavailable, but if the situation improves the user may be able to
use it more quickly.

How quickly does the circuit break now? It depends on whether the user
is using Chrome or Firefox and whether the errors encountered are
retriable or non-retriable:

- Retriable errors (e.g. connection errors, 502/503/504, etc.) will, as
  the name implies, be automatically retried. If these errors are
  continually produced, the circuit will break very quickly (if the
  extension is restarted, then it will break immediately).
- Non-retriable errors (e.g. 4xx errors) do not get automatically
  retried, so it takes longer for the circuit to break (if the extension
  is restarted, on average it will take about 1 minute).
- Note that Chrome implements "anti-DDoS throttling logic" which means
  that some non-retriable errors will turn into retriable errors. In
  this situation the circuit breaks faster than it would on Firefox.