refactor: Refactor integrationm between UserOperationController and TransactionController
The UserOperationController now triggers transaction added/updated events in
the TransactionController by using the new `emulateNewTransaction` and
`emulateTransactionUpdate` added in v60.10.0 of
`@metamask/transaction-controller`. These event handlers have also been
migrated out of `metamask-controller.js` and into the init module, which is
where these sort of integrations belong.

This helps unblock #31843. The old integration was incompatible with the new
messenger, which does not allow publishing an event outside of its namespace.